backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
firebase-admin>=6.2.0
flask-limiter>=3.5.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
google-generativeai>=0.3.1
sympy>=1.12
requests==2.31.0
Werkzeug==3.0.1
gunicorn==21.2.0
pytest==7.4.3
black==23.11.0
mypy==1.7.1
cachetools>=5.3.2
pylint==3.0.2
python-dotenv>=1.0.0
python-json-logger==2.0.7
cryptography==41.0.5
PyJWT==2.8.0
cachetools==5.3.2
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
grpcio==1.59.3
grpcio-status==1.59.3
protobuf==4.25.1
